requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-tailwind==0.5.0
Django==4.1.7
django-browser-reload==1.7.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-environ==0.10.0
django-jazzmin==2.6.0
django-tailwind==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
PyJWT==2.6.0
pytz==2023.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
